Initial commit of functionality
This provides the initial set of RSpec Shared Examples which Gumboot
will provide to AAF applications along with a dummy Rails app
implementing the proposed models and a set of local specs to ensure
all shared examples are functioning.

At this point no generators exist instead I've opted for documenting
the set of expectations for an application using Gumboot in the
readme. To discuss +/- of this approach with other developers.